Added user control to modify email API call and added test case

+++ backend/ParkMe/src/main/java/com/spm/ParkMe/controllers/ModificationController.java
@@ -1,9 +1,14 @@
 package com.spm.ParkMe.controllers;
 
+import java.security.Principal;
+
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -25,13 +30,17 @@ public class ModificationController {
 	private UserRepository repository;
 	
 	@PostMapping("/email")
-	public User modifyEmail(@Valid @RequestBody ChangeMailInfo mailInfo) {
-		User user = repository.findByUsername(mailInfo.getCurrentEmail()).orElseThrow(() -> new UsernameNotFoundException("The provided current email is wrong."));
-		//delete user from db
-		repository.delete(user);
-		System.out.println(user.getEmail());
-		user.setEmail(mailInfo.getNewEmail());
-		repository.save(user);
-		return user;
+	public ResponseEntity<User> modifyEmail(Authentication authentication, @Valid @RequestBody ChangeMailInfo mailInfo) {
+		String authenticatedUsername = authentication.getName();
+		if(authenticatedUsername.equals(mailInfo.getCurrentEmail())) {
+			User user = repository.findByUsername(mailInfo.getCurrentEmail()).orElseThrow(() -> new UsernameNotFoundException("The provided current email is wrong."));
+			//delete user from db
+			repository.delete(user);
+			user.setUsername(mailInfo.getNewEmail());
+			user.setEmail(mailInfo.getNewEmail());
+			repository.save(user);
+			return ResponseEntity.ok(user);
+		}
+		return new ResponseEntity<User>(HttpStatus.UNAUTHORIZED);
 	}
 }
